Save/Load. Ship movement. Ship shooting.

+++ MyGame/SpaceGame/Position3D.cpp
@@ -2,7 +2,7 @@
 
 Position3D::Position3D() : Position2D(0, 0), _z(0){}
 
-Position3D::Position3D(int x, int y, int z) : Position2D(x, y), _z(z)
+Position3D::Position3D(float x, float y, float z) : Position2D(x, y), _z(z)
 {
 }
 
@@ -12,6 +12,6 @@ Position3D::~Position3D()
 }
 
 
-int Position3D::getZ() { return _z; }
+float Position3D::getZ() { return _z; }
 
-void Position3D::setZ(int z) { _z = z; }
+void Position3D::setZ(float z) { _z = z; }
