+++ MyGame/SpaceGame/Region.cpp
@@ -6,7 +6,7 @@ _offSetX(0), _offSetY(0), _width(0), _height(0)
 {
 }
 
-Region::Region(int offsetY, int offsetX, int width, int height):
+Region::Region(float offsetY, float offsetX, float width, float height) :
 _offSetX(offsetX), _offSetY(offsetY), _width(width), _height(height)
 {
 }
@@ -16,13 +16,13 @@ Region::~Region()
 {
 }
 
-int Region::getOffSetY(){ return _offSetY; }
-int Region::getOffSetX(){ return _offSetX; }
-int Region::getWidth(){ return _width; }
-int Region::getHeight(){ return _height; }
+float Region::getOffSetY(){ return _offSetY; }
+float Region::getOffSetX(){ return _offSetX; }
+float Region::getWidth(){ return _width; }
+float Region::getHeight(){ return _height; }
 
 //Setters
-void Region::setOffSetY(int offsetY) { _offSetY = offsetY; }
-void Region::setOffSetX(int offsetX){ _offSetX = offsetX; }
-void Region::setWidth(int width){ _width = width; }
-void Region::setHeight(int height){ _height = height; }
+void Region::setOffSetY(float offsetY) { _offSetY = offsetY; }
+void Region::setOffSetX(float offsetX){ _offSetX = offsetX; }
+void Region::setWidth(float width){ _width = width; }
+void Region::setHeight(float height){ _height = height; }
